requirements.txt
aiogram~=3.17.0
beautifulsoup4~=4.12.0
pillow~=11.1.0
python-dotenv~=1.0.0
requests~=2.32.0
httpx~=0.28.0
celery~=5.4.0
cairosvg~=2.7.0
redis~=5.2.0
Flask~=3.1.0
opencv-python~=4.11.0.0
pytest~=8.3.4